fix: Separate mock setup
