+++ src/web/src/globals.d.ts
@@ -1,0 +1,1 @@
+declare const IS_DEV_BUILD: true | false;
