+++ src/web/src/widgets/actor-detail-widget.tsx
@@ -1,21 +1,7 @@
 import { ActorSearchDetail } from "../pages/ActorSearch/ActorSearchDetail";
-import { setupMockOpenAi } from "../utils/mock-openai";
-import { MOCK_ACTOR_DETAILS_RESPONSE } from "../utils/mock-actor-details";
 import { renderWidget } from "../utils/init-widget";
 import { useWidgetProps } from "../hooks/use-widget-props";
 import type { ActorDetails } from "../types";
-
-const shouldEnableMocks = typeof window !== "undefined" && !window.openai;
-
-if (shouldEnableMocks) {
-    // Set up a mock window.openai for local development
-    setupMockOpenAi({
-        toolOutput: {
-            details: MOCK_ACTOR_DETAILS_RESPONSE.structuredContent.actorDetails,
-        },
-        initialWidgetState: {},
-    });
-}
 
 interface WidgetToolOutput extends Record<string, unknown> {
     details?: ActorDetails;
@@ -32,4 +18,10 @@ const ActorDetailWrapper = () => {
     return <ActorSearchDetail details={details} />;
 };
 
-renderWidget(ActorDetailWrapper);
+(async () => {
+    if (IS_DEV_BUILD) {
+        const { setupActorDetailWidgetDev } = await import("./actor-detail-widget.dev");
+        setupActorDetailWidgetDev();
+    }
+    renderWidget(ActorDetailWrapper);
+})();
